chore: 🔧 move calendar utils to utils directory
